refactor: 冗長化改善のための背景黄色検出クラスの分離
